float merge

+++ src/main/java/com/graphaware/reco/neo4j/engine/Neo4jPrecomputedEngine.java
@@ -24,7 +24,7 @@ import org.neo4j.graphdb.Node;
 import org.neo4j.graphdb.Relationship;
 import org.neo4j.graphdb.RelationshipType;
 
-import static com.graphaware.common.util.PropertyContainerUtils.getInt;
+import static com.graphaware.common.util.PropertyContainerUtils.getFloat;
 import static org.neo4j.graphdb.Direction.OUTGOING;
 
 /**
@@ -63,7 +63,7 @@ public class Neo4jPrecomputedEngine extends PrecomputedEngine<Node, Node, Relati
         Score score = new Score();
 
         for (String scoreName : relationship.getPropertyKeys()) {
-            score.add(scoreName, getInt(relationship, scoreName, 0));
+            score.add(scoreName, getFloat(relationship, scoreName, 0));
         }
 
         recommendations.add(recommendation, score);
